#24 #19 changed project structure and handled validation
