+++ project-module/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.data.mongodb.uri=mongodb://AdithyaHS:adash569@ds113855.mlab.com:13855/project_add
+spring.data.mongodb.database=project_add
